+++ backstage-plugins/plugins/aws-apps/dev/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { createDevApp } from '@backstage/dev-utils';
+import { opaPlugin, AppCatalogPage } from '../src/plugin';
+
+createDevApp()
+  .registerPlugin(opaPlugin)
+  .addPage({
+    element: <AppCatalogPage kind={'all'} />,
+    title: 'Root Page',
+    path: '/aws-apps',
+  })
+  .render();