+++ backstage-plugins/plugins/aws-apps/src/pages/AwsECSEnvironmentProviderPage/AwsECSEnvironmentProviderPage.tsx
@@ -4,7 +4,7 @@
 import { Entity } from '@backstage/catalog-model';
 import { EntityAboutCard, EntityLayout, EntityLinksCard } from '@backstage/plugin-catalog';
 import { EntityCatalogGraphCard } from '@backstage/plugin-catalog-graph';
-import { isGithubActionsAvailable } from '@backstage/plugin-github-actions';
+import { isGithubActionsAvailable } from '@backstage-community/plugin-github-actions';
 import { EntityTechdocsContent } from '@backstage/plugin-techdocs';
 import { ReportIssue } from '@backstage/plugin-techdocs-module-addons-contrib';
 import { TechDocsAddons } from '@backstage/plugin-techdocs-react';
@@ -35,12 +35,12 @@ export function AwsECSEnvironmentProviderPage(/* {children}: AwsEnvironmentProvi
             <EntityCatalogGraphCard variant="gridItem" height={400} showArrowHeads />
           </Grid>
           <Grid item md={6}>
-            <EntityProviderInfoCard/>
+            <EntityProviderInfoCard />
           </Grid>
           <Grid item md={6} xs={12}>
             <EntityLinksCard />
           </Grid>
-        
+
         </Grid>
       </EntityLayout.Route>
       <EntityLayout.Route path="/ci-cd" title="CI/CD" if={isCicdApplicable}>
@@ -54,9 +54,9 @@ export function AwsECSEnvironmentProviderPage(/* {children}: AwsEnvironmentProvi
         </EntityTechdocsContent>
       </EntityLayout.Route>
       <EntityLayout.Route path="/management" title="Management">
-         <Grid item md={6} xs={12}>
-            <EntityDeleteProviderCard />
-          </Grid>
+        <Grid item md={6} xs={12}>
+          <EntityDeleteProviderCard />
+        </Grid>
       </EntityLayout.Route>
     </EntityLayout>
   );