+++ backstage-plugins/plugins/aws-apps/src/pages/AwsEKSEnvironmentProviderPage/AwsEKSEnvironmentProviderPage.tsx
@@ -1,7 +1,7 @@
 import { Entity } from '@backstage/catalog-model';
 import { EntityAboutCard, EntityLayout, EntityLinksCard } from '@backstage/plugin-catalog';
 import { EntityCatalogGraphCard } from '@backstage/plugin-catalog-graph';
-import { isGithubActionsAvailable } from '@backstage/plugin-github-actions';
+import { isGithubActionsAvailable } from '@backstage-community/plugin-github-actions';
 import { EntityTechdocsContent } from '@backstage/plugin-techdocs';
 import { ReportIssue } from '@backstage/plugin-techdocs-module-addons-contrib';
 import { TechDocsAddons } from '@backstage/plugin-techdocs-react';
@@ -32,7 +32,7 @@ export function AwsEKSEnvironmentProviderPage(/* {children}: AwsEnvironmentProvi
             <EntityCatalogGraphCard variant="gridItem" height={400} showArrowHeads />
           </Grid>
           <Grid item md={6}>
-            <EntityProviderInfoCard/>
+            <EntityProviderInfoCard />
           </Grid>
           <Grid item md={6} xs={12}>
             <EntityLinksCard />
@@ -50,9 +50,9 @@ export function AwsEKSEnvironmentProviderPage(/* {children}: AwsEnvironmentProvi
         </EntityTechdocsContent>
       </EntityLayout.Route>
       <EntityLayout.Route path="/management" title="Management">
-         <Grid item md={6} xs={12}>
-            <EntityDeleteProviderCard />
-          </Grid>
+        <Grid item md={6} xs={12}>
+          <EntityDeleteProviderCard />
+        </Grid>
       </EntityLayout.Route>
     </EntityLayout>
   );