+++ backstage-plugins/plugins/scaffolder-backend-module-aws-apps/src/actions/get-platform-parameters/get-platform-parameters.ts
@@ -11,7 +11,7 @@ const ID = 'opa:get-platform-parameters';
 
 const examples = [
   {
-    description: 'Retrieve AWS SSM parameter values for the OPA on AWS platform so that their values can be used by other template actions',
+    description: 'Retrieve AWS SSM parameter values for the Harmonix on AWS platform so that their values can be used by other template actions',
     example: yaml.stringify({
       steps: [
         {