+++ build-script/local-runners.sh
@@ -7,7 +7,7 @@ source $scriptDir/helpers.sh
 
 start_local() {
 	echo "Starting a local PostgreSQL container"
-	docker compose -f $appRootDir/config/docker-compose.yml up --build --detach
+	docker compose -p "local-harmonix" -f $appRootDir/config/docker-compose.yml up --build --detach
 	echo "Starting the backstage app"
 	set -a && source $appRootDir/config/.env && set +a && yarn --cwd $backstageDir dev
 	# yarn --cwd $backstageDir dev